requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arxiv==2.2.0
attrs==25.1.0
beautifulsoup4==4.13.3
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cssselect2==0.8.0
dataclasses-json==0.6.7
decorator==5.2.1
-e git+https://github.com/bytedance/deer-flow.git@32d8e514e1ed17d6dd3fd977d2816a898bdb0a89#egg=deer_flow
distro==1.9.0
duckduckgo_search==8.0.0
fastapi==0.115.11
feedparser==6.0.11
filelock==3.18.0
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.5.0
inquirerpy==0.3.4
Jinja2==3.1.6
jiter==0.8.2
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.65
langchain-deepseek==0.1.3
langchain-experimental==0.3.4
langchain-mcp-adapters==0.0.9
langchain-openai==0.3.22
langchain-text-splitters==0.3.6
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.45
litellm==1.63.11
lxml==5.3.1
Markdown==3.8.2
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.6.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==2.2.3
openai==1.86.0
orjson==3.10.15
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
peewee==3.17.9
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.3.6
primp==0.14.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.5
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.9.9
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydyf==0.11.0
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
readabilipy==0.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retry==0.9.2
rpds-py==0.23.1
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.38
sse-starlette==2.1.3
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.0
volcengine==1.0.191
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.54
zipp==3.21.0
zopfli==0.2.3.post1
zstandard==0.23.0
